[GREEN] GetFiguresByClient, changed radius type from int to decimal

+++ ObligatorioDA1/Domain/Figures/Sphere.cs
@@ -2,6 +2,6 @@
 {
     public class Sphere : Figure
     {
-        public int Radius { get; set; }
+        public decimal Radius { get; set; }
     }
 }
